+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Manage `weight` props on each elements of `linkList` component.
 - Feature: Add `weight` props on `LinkList` component.
 
 ## [23.0.0] - 2018-08-27